+++ lib/hit.dart
@@ -1,8 +1,8 @@
 library;
 
-export 'src/hit_defer.dart' show Hit;
+export 'src/hit_defer.dart' show HitDefer;
 export 'src/hit_layer.dart' show HitLayer;
-export 'src/hit_link.dart' show HitLink, HitDeferRegistration;
+export 'src/hit_link.dart' show HitLink, HitDeferRegistration, HitDeferPaint;
 export 'src/hit_scope.dart'
     show
         HitScope,